### 2017-02-12 #180 @arfon
:wave: @eteq - would you be willing to review this submission for JOSS? You can learn more about what that entails here: http://joss.theoj.org/about#reviewer_guidelines
