### 2017-02-13 #182 @whedon
[REVIEW]: Cine: A solar-pumped fluorescence model for cometary atmospheres
**Submitting author:** @migueldvb (<a href="http://orcid.org/0000-0002-0455-9384">Miguel de Val-Borro</a>)
**Repository:** <a href="https://github.com/migueldvb/cine" target ="_blank">https://github.com/migueldvb/cine</a>
**Version:** v0.2.1
