### 2017-02-21 migueldvb/cine#1 @eteq
No orcid for Steven B. Charnley ?
While reviewing the JOSS paper for openjournals/joss-reviews#182, I notice there's an author with no orcid (Steven B. Charnley).  Is this intended (i.e., he has no orcid), or an oversight?
