### 2017-02-21 #182 @eteq
Process question (maybe for @arfon, maybe for @whedon?) - should there be a PDF of the paper in the author's repo, or just the ``.md`` and ``.bib`` files (which are there already)?
