### 2017-02-22 #182 @arfon
> Process question (maybe for @arfon, maybe for @whedon?) - should there be a PDF of the paper in the author's repo, or just the .md and .bib files (which are there already)?

`.md` and `.bib` files are fine. We generally compile these at the end into a PDF.
