### 2017-02-22 migueldvb/cine#2 @eteq
Setuptools should be a requirement/bootstrapped
If I try to do a ``python setup.py build`` on a minimal python setup, I get ``ImportError: No module named setuptools``.  There are two solutions I know of:

1. use https://bootstrap.pypa.io/ez_setup.py as a bootstrap script which auto-installs setuptools.
2. add "pip install setuptools" to the setup instructions (and maybe catch the import error in your setup.py)

I'd say just do 1, except that its notes say that approach is deprecated.  So either of these work.
