### 2017-02-22 d792621
Fallback import to use distutils
This will work when setuptools is not available
