### 2017-02-22 migueldvb/cine#2 @migueldvb
That is a good point. I have added a fallback import of setup from distutils. The `setup.py` script should work now with both setuptools and distutils.
