### 2017-02-22 6fd69cb
add scipy pandas and astroquery as requirements
