### 2017-02-22 migueldvb/cine#4 @eteq
Python 2 not supported?
If I try to import `cine` from python 2.7, I get a ``SyntaxError: only named arguments may follow *expression``... But the ``setup.py`` implies that 2.x is supposed to be supported.

I'm agnostic whether you want to support py2 or not, but it should be stated in the readme (and the ``setup.py`` should be consistent what major version(s) of python are supported.
