### 2017-02-22 migueldvb/cine#5 @eteq
Clarify if this is meant to be command-line only or also python-useable
From the first part of  the README I get the impression that I'm supposed to use this as a Python library.  But then the example suggests this is meant to be a command-line code.  That should be clarified.

*Also*, if it's supposed to be used as a library, there needs to be more documentation (can say more if that's the case).
