### 2017-02-22 #182 @eteq
Alright, I've finished my initial review.  I've created migueldvb/cine#1,  migueldvb/cine#2, migueldvb/cine#4, migueldvb/cine#5 with detailed comments that need addressing.

One higher-level question (and the reason for the one un-checked item above) is the lack of API documentation.  I think this *might* be intentional (see migueldvb/cine#5), but if not, that needs to be rectified.

One other procedural question for @arfon: in the process of doing this review I created a PR that @migueldvb merged (it was easier just to do that).  Nominally I suppose that would now make me a co-author given that I have a commit in the repo.  But of course that would look like a weird conflict-of-interest.  Is there a policy for how this should work?  (I'm perfectly fine *not* being an author, as it's a very minor change...  But this is also sort of an academic curiosity.)
