### 2017-02-22 migueldvb/cine#4 @eteq
(This is part of the review in openjournals/joss-reviews#182)
