### 2017-02-22 migueldvb/cine#5 @eteq
(This is part of the review in openjournals/joss-reviews#182)
