### 2017-02-23 #182 @arfon
> One other procedural question for @arfon: in the process of doing this review I created a PR that @migueldvb merged (it was easier just to do that). Nominally I suppose that would now make me a co-author given that I have a commit in the repo. But of course that would look like a weird conflict-of-interest. Is there a policy for how this should work? (I'm perfectly fine not being an author, as it's a very minor change... But this is also sort of an academic curiosity.)

Great question @eteq. I'm in the same situation for many of the submissions as I make small edits to the `paper.md`files. Where I'm at on this:

- I think that minor contributions made during the process of review are OK but should not result in the reviewing becoming an author
- More major contributions which would likely warrant authorship are also OK but in this scenario then I think we'd want to find a second reviewer

Does that sound reasonable?
