+++ requirements.txt
@@ -1,0 +1,4 @@
+numpy
+scipy
+pandas
+astroquery
