### 2017-02-23 #182 @eteq
Yep, sounds good @arfon - in that case no issue here as the PR I sent definitely falls in the "minor" category.
