### 2017-02-23 bffbce8
Workaround to upgrade pip in travis environment
