+++ .travis.yml
@@ -4,6 +4,9 @@ python:
   - "3.4"
   - "3.5"
   - "3.6"
+# workaround to get a recent pip version that is able to handle the order of
+# dependencies
+before_install: "pip install -U pip"
 # command to install dependencies
 install: "pip install -r requirements.txt"
 # command to run tests
