### 2017-02-23 2e0194d
Use named arguments in call to download_hitran function
This fixes the build in python 2.7 and 3.4
