### 2017-02-23 migueldvb/cine#4 @migueldvb
Thanks for catching this issue. It looks like this is the only thing that prevents the code from working in python 2.7 so I will fix this and keep support for 2.7 in setup.py.

There is now travis continuous integration set up in the repository to build and test the software after every commit in python 2.7, 3.4, 3.5 and 3.6 that should find these problems. I have also added the badge that shows the status of the build in the README file.
