+++ .travis.yml
@@ -6,8 +6,14 @@ python:
   - "3.6"
 # workaround to get a recent pip version that is able to handle the order of
 # dependencies
-before_install: "pip install -U pip"
+before_install:
+  - pip install -U pip
 # command to install dependencies
-install: "pip install -r requirements.txt"
+install:
+  - pip install -r requirements.txt
+  - pip install python-coveralls
 # command to run tests
-script: python setup.py test
+script:
+  - nosetests --with-coverage --cover-package=cine
+after_success:
+  - coveralls
