### 2017-02-25 #182 @eteq
@migueldvb - just to clarify in case you missed it above: migueldvb/cine#1 and migueldvb/cine#5 are currently awaiting response. Depending on the answer to migueldvb/cine#5, there *may* be additional documentation requirements, but once those issues are addressed, I think this is good to go.
