### 2017-02-25 migueldvb/cine#5 @migueldvb
At this point it is only useful to run the code using the command line tool... Nevertheless I will generate some documentation using sphinx as this may be useful to combine the fluorescence calculation in python models to be used in the LIME radiative transfer code when this functionality is available (https://github.com/lime-rt/lime/pull/33).

This should be done this coming week.
