### 2017-02-25 migueldvb/cine#1 @migueldvb
I think he plans to register with orcid soon. Now the markdown file compiles using the command `pandoc -S -o paper.pdf -V geometry:margin=1in --filter pandoc-citeproc paper.md --template /path/to/latex.template` without the orcid identifier but will be happy to add the number later.
