### 2017-02-27 migueldvb/cine#5 @eteq
OK, great, @migueldvb. In that case, the only part I think is *necessary* at this point (for the paper) is to note in the README that it's primarily intended to be a command-line tool.  (But your idea of adding docs w/sphinx is a nice bonus!)
