### 2017-02-27 migueldvb/cine#1 @eteq
Will leave it open for whenever he does register, but I presume not critical for JOSS acceptance
