### 2017-02-27 dbd9295
rewrite README to make clear this is a command-line tool
