### 2017-02-27 migueldvb/cine#5 @migueldvb
I have reworded the README file to make clear that this is supposed to be used as a command-line script for the moment. At this point this is the only useful way to generate pumping rates as input for molecular excitation and radiative transfer, but will be glad to add the docs for helping with integration with other codes.
