### 2017-02-27 #182 @migueldvb
@eteq thank you for the review.

I have clarified that the code is supposed to be run as a script in the README file. In the future it could be useful to be imported into python when defining the input model using python in the excitation code LIME is available. There is some work done in this direction but it has not yet been merged into LIME and this may take some time...

The missing orcid in the paper is because the co-author does not have an identifier yet although he is interested in getting one. I will leave the issue open as a reminder but perhaps we should not wait for this if all the other requirements are met.
