### 2017-02-28 migueldvb/cine#5 @eteq
OK, cool.  I think this satisfies the review requirement.

Do you want to leave this open to remind you to add the integration docs, @migueldvb, or should we close this and make a new one?
