### 2017-02-28 migueldvb/cine#5 @migueldvb
Thanks, it is OK to leave this issue open.
