### 2017-03-06 #182 @eteq
Alright, great @migueldvb.

@arfon, the author has addressed my concerns, so I think I can recommend this paper for acceptance.  There is one question remaining, though: as indicated in migueldvb/cine#1, one of the co-authors doesn't have an orcid yet, but @migueldvb thinks he might get one later.  Is it possible to update the author list after the facts and have whatever metadata is relevant be updated, or is the orcid a now-or-nothing thing?
