### 2017-03-06 #182 @arfon
> Is it possible to update the author list after the facts and have whatever metadata is relevant be updated, or is the orcid a now-or-nothing thing?

It's now or never I'm afraid. We only update paper metadata if there's been an error with a submission.
