### 2017-03-06 #182 @arfon
> If the paper can be accepted as it is with a missing orcid for one author, I would go ahead without it and close the issue in the repo.

It can.
