### 2017-03-06 migueldvb/cine#1 @migueldvb
Closing this as it will not be fixed before acceptance.
