### 2017-03-06 #182 @migueldvb
@arfon I have made a DOI for the v0.2.1 release with Zenodo:
https://doi.org/10.5281/zenodo.345982

The title in the markdown paper file is changed to be more especific, 'Cine: line excitation by infrared fluorescence in cometary atmospheres', could you update it in the submission?
