### 2017-03-06 1c4dc10
Update version number in setup.py
