### 2017-03-07 #182 @arfon
Many thanks for the excellent review here @eteq ✨

@migueldvb - your paper is now accepted into JOSS and the DOI is http://dx.doi.org/10.21105/joss.00182 ⚡️ 🚀 💥
